data_pipeline/requirements.txt
pandas==2.2.1
openpyxl==3.1.2
polars==0.20.15
psycopg2-binary==2.9.9
pgvector==0.2.4
Django==5.0.3
paramiko==3.4.0
python-dotenv==1.0.1
sentence-transformers==2.6.1
langchain==0.1.16
langgraph==0.0.38
langchain-openai==0.1.3
pytest==8.1.1
faker==24.3.0
# ftplib is part of Python standard library — no install needed
